\name{name-methods}
\docType{methods}
\alias{name}
\alias{name-methods}
\alias{name,flowFP-method}
\alias{name,flowFPModel-method}
\alias{name,flowFPPlex-method}
\title{ Get Model Name}
\description{
 These methods return the model name(s) for an object.
}
\section{Methods}{
\code{name(object)}
\describe{

\item{object = "flowFP"}{Returns the name of the model used to create these 
                         fingerprints.}

\item{object = "flowFPModel"}{Returns the model name.}

\item{object = "flowFPPlex"}{Returns all of the model names found in the plex.}
}}
\keyword{methods}
