\name{nInstances-methods}
\docType{methods}
\alias{nInstances}
\alias{nInstances-methods}
\alias{nInstances,flowFP-method}
\alias{nInstances,flowFPPlex-method}
\title{Returns the number of instances in an object.}
\description{
 Theses methods return the number of \emph{flowFrame}s that where used to create a fingerprint or plex.
}
\section{Methods}{
\code{nInstances(object)}
\describe{

\item{object = "flowFP"}{The number of instances in a fingerprint.}

\item{object = "flowFPPlex"}{The total number of instances in a plex fingerprint.}
}}
\keyword{methods}
