\name{parameters-methods}
\docType{methods}
\alias{parameters-methods}
\alias{parameters,flowFP-method}
\alias{parameters,flowFPModel-method}
\title{ Return the parameters considered for fingerprinting}
\description{
 These methods return the flow parameters that were taken into
consideration at \link{flowFPModel} construction.
}
\section{Methods}{
\code{parameters(object)}
\describe{

\item{object = "flowFP"}{Returns the parameters considered when creating
                         fingerprint.}

\item{object = "flowFPModel"}{Returns the parameters that will be used to create 
                              a fingerprint.}

}}
\keyword{methods}
