\name{sampleClasses-methods}
\docType{methods}
\alias{sampleClasses}
\alias{sampleClasses<-}
\alias{sampleClasses<-,flowFP-method}
\alias{sampleClasses<-,flowFPPlex-method}
\alias{sampleClasses-methods}
\alias{sampleClasses,flowFP-method}
\alias{sampleClasses,flowFPPlex-method}
\title{ Gets/Sets the sample classes for a fingerprint.}
\description{
 These methods get or set the sample classes for a fingerprint or a set of fingerprints.
}
\section{Methods}{
\code{sampleClasses(object)}

\describe{

\item{object = "flowFP"}{ Returns the sampleClass factor object found in a flowFP }
\item{object = "flowFPPlex"}{ Returns the sampleClass factor object found in a flowFPPlex }
}}

\keyword{methods}
