\name{sampleNames-methods}
\docType{methods}
\alias{sampleNames}
\alias{sampleNames-methods}
\alias{sampleNames,flowFP-method}
\alias{sampleNames,flowFPPlex-method}
\title{ Methods to return sample names from a flowFP object.}
\description{
	These methods get the sample names for a fingerprint or a set of fingerprints.
}
\section{Methods}{
\code{sampleNames(object)}
\describe{

\item{object = "flowFP"}{ returns the sample names in a flowFP object.}
\item{object = "flowFPPlex"}{ returns the sample names in a flowFPPlex object.}

}}
\keyword{methods}
