\name{show-methods}
\docType{methods}
\alias{show-methods}
\alias{show,flowFP-method}
\alias{show,flowFPModel-method}
\alias{show,flowFPPlex-method}
\title{ Methods to view flowFP objects}
\description{
 These methods give basic information about the contents of a \emph{flowFP} object.
}
\section{Methods}{
\code{show(object)}
\describe{

\item{object = "flowFP"}{Show the sample names and model information for a flowFP. }

\item{object = "flowFPModel"}{Shows the samples used in creating the model. }

\item{object = "flowFPPlex"}{Shows all of the sample named and all of the models
                             found in the plex.}
}}
\keyword{methods}
