\name{summary-methods}
\docType{methods}
\alias{summary-methods}
\alias{summary,flowFP-method}
\title{ Summerizes a flowFP object.}
\description{
 These methods give a summary of the flowFP object.
}
\section{Methods}{
\code{summary(object)}
\describe{

\item{object = "flowFP"}{Show the per-instance min/max/mean counts in a\code{flowFP}. }

\item{object = "flowFPModel"}{Shows length, class and mode info for a \code{flowFPModel}. }

\item{object = "flowFPPlex"}{Shows length, class and mode info for a \code{flowFPPlex}.}

}}
\keyword{methods}
